% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qk4d.R
\name{qk4d}
\alias{qk4d}
\title{qunatile of k4d}
\usage{
qk4d(p, loc = 0, scale = 1, shape1 = 0.1, shape2 = 0.1)
}
\arguments{
\item{p}{Vector of quantiles.}

\item{loc, scale, shape1, shape2}{Location, scale, and shape parameters. Can be vectors, but the lengths must be appropriate.}
}
\value{
quantile function for the 1-GGD distribution
}
\description{
qunatile of k4d
}
\examples{
\dontrun{
qk4d(0.5, loc = 0, scale=1, shape1 = 0.1, shape2 = 0.1)
}
}
