% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qld.R
\name{qld}
\alias{qld}
\title{qunatile of logistic}
\usage{
qld(p, loc = 0, scale = 1)
}
\arguments{
\item{p}{Vector of quantiles.}

\item{loc, scale}{Location, scale parameters. Can be vectors, but the lengths must be appropriate.}
}
\value{
quantile function for the 1-LD distribution
}
\description{
qunatile of logistic
}
\examples{
\dontrun{
qld(0.5, loc = 0, scale=1)
}
}
