% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgd.fit.R
\name{rgd.fit}
\alias{rgd.fit}
\title{Fit the Gumbel Distribution to r-Largest Order Statistics}
\usage{
rgd.fit(
  xdat,
  r = NULL,
  ydat = NULL,
  mul = NULL,
  sigl = NULL,
  mulink = identity,
  siglink = identity,
  num_inits = 100,
  muinit = NULL,
  siginit = NULL,
  show = TRUE,
  method = "Nelder-Mead",
  maxit = 10000,
  ...
)
}
\arguments{
\item{xdat}{A numeric vector, matrix, or data frame of observations.
Each row should contain decreasing order statistics for a given year
or block. The first column therefore contains block maxima. Only the
first \code{r} columns are used in the fitted model. If \code{r} is
\code{NULL}, all available columns are used. If some rows contain fewer
order statistics than others, missing values should be appended at the
end of the corresponding rows.}

\item{r}{The number of largest order statistics to use in the fitted model.
If \code{NULL}, all columns of \code{xdat} are used.}

\item{ydat}{A matrix or data frame of covariates for non-stationary modelling
of the parameters, or \code{NULL} for a stationary model. The number of rows
must match the number of rows of \code{xdat}.}

\item{mul, sigl}{Integer vectors indicating which columns of \code{ydat} are
used as covariates for the location and scale parameters, respectively.
Use \code{NULL} for stationary parameters.}

\item{mulink, siglink}{Inverse link functions for the location and scale
parameters, respectively.}

\item{num_inits}{The number of initial parameter sets used in the optimization.}

\item{muinit, siginit}{Numeric vectors giving initial values for the location
and scale parameters. If \code{NULL}, default initial values based on
L-moments are used.}

\item{show}{Logical. If \code{TRUE}, details of the fitted model are printed.}

\item{method}{Optimization method passed to \code{\link{optim}} for stationary fits.}

\item{maxit}{Maximum number of iterations for \code{\link{optim}}.}

\item{...}{Additional control arguments passed to the optimizer.}
}
\value{
A list with components including:
\item{trans}{Logical; \code{TRUE} if a non-stationary model is fitted.}
\item{model}{A list containing \code{mul} and \code{sigl}.}
\item{link}{A character string describing the inverse link functions.}
\item{conv}{The convergence code returned by the optimizer. A value of 0
indicates successful convergence for \code{optim}.}
\item{nllh}{The negative log-likelihood evaluated at the fitted parameters.}
\item{data}{The data used in the fit.}
\item{mle}{The maximum likelihood estimates.}
\item{cov}{The estimated covariance matrix.}
\item{se}{The estimated standard errors.}
\item{vals}{A matrix containing fitted values of the location and scale
parameters at each observation.}
\item{r}{The number of order statistics used in the fitted model.}
}
\description{
Fits the Gumbel distribution to \eqn{r}-largest order statistics using
maximum likelihood estimation. Stationary and non-stationary models are
supported through generalized linear modelling of the location and scale
parameters.
}
\examples{
x <- rgdr(n = 50, r = 2, loc = 10, scale = 2)
fit <- rgd.fit(x$rmat)
fit$r
fit$mle
}
\references{
Coles, S. (2001).
An Introduction to Statistical Modeling of Extreme Values.
Springer.

Shin, Y., & Park, J.-S. (2025).
Generalized Gumbel model for r-largest order statistics with application
to peak streamflow.
\emph{Scientific Reports}.
\doi{10.1038/s41598-024-83273-y}
}
\seealso{
\code{\link{optim}}
}
