% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgd.prof.R
\name{rgd.prof}
\alias{rgd.prof}
\alias{rgd.Prof}
\title{profile likelihood for rgd}
\usage{
rgd.Prof(z, m, xlow, xup, conf = 0.95, nint = 100)
}
\arguments{
\item{z}{An object returned by \code{rggdfit}. The object should represent a stationary model.}

\item{m}{The return level (i.e.\ the profile likelihood is for the value that is exceeded with probability 1/\code{m}).}

\item{xlow, xup}{The least and greatest value at which to evaluate the profile likelihood.}

\item{conf}{The confidence coefficient of the plotted profile confidence interval.}

\item{nint}{The number of points at which the profile likelihood is evaluated.}
}
\value{
plot of the profile likelihood is produced, with a horizontal line representing a profile confidence interval with confidence coefficient \code{conf}.
}
\description{
profile likelihood for rgd
}
\examples{
data(bangkok)
\dontrun{
z<-rgd.fit(bangkok)
rgd.Prof(z,100,xlow=180,xup=230)
}
}
