% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgdr.R
\name{rgdr}
\alias{rgdr}
\title{random sample generation for rgd}
\usage{
rgdr(n, r, loc = 0, scale = 1)
}
\arguments{
\item{n}{Number of observations}

\item{r}{Number of order statistics for each observation.}

\item{loc, scale}{Location, scale. Can be vectors, but the lengths must be appropriate.}
}
\value{
Random number generation
}
\description{
random sample generation for rgd
}
\examples{
\dontrun{
rgdr(50, 5, loc = 0, scale=1)
}
}
