% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rggd.ed.R
\name{rggd.ed}
\alias{rggd.ed}
\alias{rggdEd}
\title{Entropy difference for rGGD}
\usage{
rggdEd(data)
}
\arguments{
\item{data}{Data should be contain n rows, each a rGGD observation}
}
\value{
list containing the following components. A subset of these components are printed after the fit.  \code{statistic}, \code{p.value}, \code{theta} are always printed.#'
}
\description{
Entropy difference for rGGD
}
\examples{
\dontrun{
data(bangkok)
rggdEd(bangkok)
}
}
