% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rggd.edtest.R
\name{rggd.edtest}
\alias{rggd.edtest}
\alias{rggdEdtest}
\title{Entropy difference test for rGGD}
\usage{
rggdEdtest(data)
}
\arguments{
\item{data}{Data should be contain n rows, each a GGDr observation.}
}
\value{
{
Function returns a dataframe containing the test statistics, estimates, and p-value results of the sequential tests.
\item{r}{Value of r to be tested.}
\item{p.values}{Raw p-values from the individual tests at each value of r.}
\item{ForwardStop}{Transformed p-values according to the ForwardStop stopping rule.}
\item{StrongStop}{Transformed p-values according to the StrongStop stopping rule.}
\item{statistic}{Returned test statistics of each individual test.}
\item{est.loc}{Estimated location parameter for the given r.}
\item{est.scale}{Estimated scale parameter for the given r.}
\item{est.shape}{Estimated shape parameter for the given r.}
}
}
\description{
Entropy difference test for rGGD
}
\examples{
\dontrun{
data(bangkok)
rggdEdtest(bangkok)
}
}
