% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rggd.lh.R
\name{rggd.lh}
\alias{rggd.lh}
\alias{rggdLh}
\title{loglikelihood for rggd}
\usage{
rggdLh(data, par)
}
\arguments{
\item{data}{Vector or matrix of observations. If x is a matrix, each row is taken to be a new observation.}

\item{par}{Location, scale, and shape parameters. Can be vectors, but the lengths must be appropriate.}
}
\value{
log-likelihood value
}
\description{
loglikelihood for rggd
}
