% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rggd.rl.R
\name{rggd.rl}
\alias{rggd.rl}
\title{return level for ggd}
\usage{
rggd.rl(z, year = c(20, 50, 100, 200), show = F)
}
\arguments{
\item{z}{An object returned by \code{rggd.fit}. The object should represent a stationary model.}

\item{year}{The return level (i.e.\ the profile likelihood is for the value that is exceeded with probability 1/\code{m}).}

\item{show}{Logical; if \code{TRUE} (the default), print details of the result.}
}
\value{
The return level and numeric vector giving the standard errors for the return level.
}
\description{
return level for ggd
}
\examples{
\dontrun{
data(bangkok)
result<-rk4d.fit(bangkok)
rggd.rl(result$mle,result$cov,100)
}
}
