% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rggdr.R
\name{rggdr}
\alias{rggdr}
\title{random sample generation for rggd}
\usage{
rggdr(n, r, loc = 0, scale = 1, shape = 0.1)
}
\arguments{
\item{n}{Number of observations}

\item{r}{Number of order statistics for each observation.}

\item{loc, scale, shape}{Location, scale, and shape parameters. Can be vectors, but the lengths must be appropriate.}
}
\value{
Random number generation
}
\description{
random sample generation for rggd
}
\examples{
\dontrun{
rggdr(50, 5, loc = 0, scale=1, shape = 0.1)
}
}
