% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rglo.ed.R
\name{rglo.ed}
\alias{rglo.ed}
\alias{rgloEd}
\title{Entropy difference for rGLO}
\usage{
rgloEd(data, par = NULL)
}
\arguments{
\item{data}{Data should be contain n rows, each a rGLO observation.}

\item{par}{Location, scale, and shape parameters. Can be vectors, but the lengths must be appropriate.}
}
\value{
list containing the following components. A subset of these components are printed after the fit.  \code{statistic}, \code{p.value}, \code{theta} are always printed.#'
}
\description{
Entropy difference for rGLO
}
\examples{
\dontrun{
data(bangkok)
rgloEd(bangkok)
}
}
