% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rglo.lh.R
\name{rglo.lh}
\alias{rglo.lh}
\alias{rgloLh}
\title{loglikelihood for rglo}
\usage{
rgloLh(data, par)
}
\arguments{
\item{data}{Vector or matrix of observations. If x is a matrix, each row is taken to be a new observation.}

\item{par}{Location, scale, and shape parameters. Can be vectors, but the lengths must be appropriate.}
}
\value{
log-likelihood value
}
\description{
loglikelihood for rglo
}
