% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rglo.rl.R
\name{rglo.rl}
\alias{rglo.rl}
\title{Return Levels for the Generalized Logistic Distribution}
\usage{
rglo.rl(z, year = c(20, 50, 100, 200), show = TRUE)
}
\arguments{
\item{z}{An object returned by \code{\link{rglo.fit}}. The fitted model
should represent a stationary model.}

\item{year}{A numeric vector of return periods for which return levels
are to be computed.}

\item{show}{Logical. If \code{TRUE}, the estimated return levels and
their standard errors are printed.}
}
\value{
The input object \code{z} with two additional components:
\item{rl}{A numeric vector of estimated return levels.}
\item{rlse}{A numeric vector of standard errors of the estimated return levels.}
}
\description{
Computes return levels and their standard errors for a stationary
generalized logistic model fitted by \code{\link{rglo.fit}}.
}
\details{
For a return period \eqn{T}, the return level is defined as the quantile
exceeded with probability \eqn{1/T}. Under the generalized logistic
distribution, the return level is
\deqn{x_T = \mu + \frac{\sigma}{\xi} \left[1 - \left(\frac{1 - 1/T}{1/T}\right)^{-\xi}\right],}
which is equivalently written in the implementation as
\deqn{x_T = \mu + \frac{\sigma}{\xi} - \frac{\sigma}{\xi}
\left(\frac{1/T}{1 - 1/T}\right)^{\xi}.}
Standard errors are obtained using the delta method.
}
\examples{
x <- rglor(n = 50, r = 3, loc = 10, scale = 2, shape = 0.1)
fit <- rglo.fit(x$rmat)
out <- rglo.rl(fit, year = c(20, 50, 100, 200))
}
\references{
Ahmad, M. I., Sinclair, C. D., and Werritty, A. (1988).
Log-logistic flood frequency analysis.
\emph{Journal of Hydrology}.
\doi{10.1016/0022-1694(88)90015-7}

Coles, S. (2001).
An Introduction to Statistical Modeling of Extreme Values.
Springer.

Shin, Y., & Park, J-S. (2024).
Generalized logistic model for r-largest order statistics with
hydrological application.
\emph{Stochastic Environmental Research and Risk Assessment}.
\doi{10.1007/s00477-023-02642-7}
}
\seealso{
\code{\link{rglo.fit}}, \code{\link{rglo.prof}}
}
