% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rglor.R
\name{rglor}
\alias{rglor}
\title{random sample generation for rglo}
\usage{
rglor(n, r, loc = 0, scale = 1, shape = 0.1)
}
\arguments{
\item{n}{Number of observations}

\item{r}{Number of order statistics for each observation.}

\item{loc, scale, shape}{Location, scale, and shape parameters. Can be vectors, but the lengths must be appropriate.}
}
\value{
Random number generation
}
\description{
random sample generation for rglo
}
\examples{
\dontrun{
rglor(50, 5, loc = 0, scale=1, shape = 0.1)
}
}
