% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rk4d.ed.R
\name{rk4d.ed}
\alias{rk4d.ed}
\alias{rk4dEd}
\title{Entropy difference for rK4d}
\usage{
rk4dEd(data)
}
\arguments{
\item{data}{Data should be contain n rows, each a rK4D observation}
}
\value{
list containing the following components. A subset of these components are printed after the fit.  \code{statistic}, \code{p.value}, \code{theta} are always printed.#'
}
\description{
Entropy difference for rK4d
}
\examples{
\dontrun{
data(bangkok)
rk4dEd(bangkok)
}
}
