% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rk4d.lh.R
\name{rk4d.lh}
\alias{rk4d.lh}
\title{loglikelihood for rk4d}
\usage{
rk4d.lh(data, par)
}
\arguments{
\item{data}{Vector or matrix of observations. If x is a matrix, each row is taken to be a new observation.}

\item{par}{Location, scale, and shape parameters. Can be vectors, but the lengths must be appropriate.}
}
\value{
log-likelihood value
}
\description{
loglikelihood for rk4d
}
