% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rk4dr.R
\name{rk4dr}
\alias{rk4dr}
\title{random sample generation for rk4d}
\usage{
rk4dr(n, r, loc = 0, scale = 1, shape1 = 0.1, shape2 = 0.1)
}
\arguments{
\item{n}{Number of observations}

\item{r}{Number of order statistics for each observation.}

\item{loc, scale, shape1, shape2}{Location, scale, and shape parameters. Can be vectors, but the lengths must be appropriate.}
}
\value{
Random number generation
}
\description{
random sample generation for rk4d
}
\examples{
\dontrun{
rk4dr(50, 5, loc = 0, scale=1, shape1 = 0.1, shape2 = 0.1)
}
}
