% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rldr.R
\name{rldr}
\alias{rldr}
\title{random sample generation for rld}
\usage{
rldr(n, r, loc = 0, scale = 1)
}
\arguments{
\item{n}{Number of observations}

\item{r}{Number of order statistics for each observation.}

\item{loc, scale}{Location, scale. Can be vectors, but the lengths must be appropriate.}
}
\value{
Random number generation
}
\description{
random sample generation for rld
}
\examples{
\dontrun{
rldr(50, 5, loc = 0, scale=1)
}
}
