% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgd.R
\name{qgd}
\alias{qgd}
\title{qunatile of gumbel}
\usage{
qgd(p, loc = 0, scale = 1)
}
\arguments{
\item{p}{Vector of quantiles.}

\item{loc, scale}{Location, scale parameters. Can be vectors, but the lengths must be appropriate.}
}
\value{
quantile function for the 1-GD distribution
}
\description{
qunatile of gumbel
}
\examples{
\dontrun{
qgd(0.5, loc = 0, scale=1)
}
}
