% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qggd.R
\name{qggd}
\alias{qggd}
\title{qunatile of ggd}
\usage{
qggd(p, loc = 0, scale = 1, shape = 0)
}
\arguments{
\item{p}{Vector of quantiles.}

\item{loc, scale, shape}{Location, scale, and shape parameters. Can be vectors, but the lengths must be appropriate.}
}
\value{
quantile function for the 1-GGD distribution
}
\description{
qunatile of ggd
}
\examples{
\dontrun{
qggd(0.5, loc = 0, scale=1, shape = 0.1)
}
}
