% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qglo.R
\name{qglo}
\alias{qglo}
\title{qunatile of glo}
\usage{
qglo(p, loc = 0, scale = 1, shape = 0)
}
\arguments{
\item{p}{Vector of quantiles.}

\item{loc, scale, shape}{Location, scale, and shape parameters. Can be vectors, but the lengths must be appropriate.}
}
\value{
quantile function for the 1-GGD distribution
}
\description{
qunatile of glo
}
\examples{
\dontrun{
qglo(0.5, loc = 0, scale=1, shape = 0.1)
}
}
