% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rglo.prof.R
\name{rglo.prof}
\alias{rglo.prof}
\alias{rglo.Prof}
\title{profile likelihood for rglo}
\usage{
rglo.Prof(z, m, xlow, xup, conf = 0.95, nint = 100)
}
\arguments{
\item{z}{An object returned by \code{rglofit}. The object should represent a stationary model.}

\item{m}{The return level (i.e.\ the profile likelihood is for the value that is exceeded with probability 1/\code{m}).}

\item{xlow, xup}{The least and greatest value at which to evaluate the profile likelihood.}

\item{conf}{The confidence coefficient of the plotted profile confidence interval.}

\item{nint}{The number of points at which the profile likelihood is evaluated.}
}
\value{
plot of the profile likelihood is produced, with a horizontal line representing a profile confidence interval with confidence coefficient \code{conf}.
}
\description{
profile likelihood for rglo
}
\examples{
data(bangkok)
\dontrun{
z<-rglo.fit(bangkok)
rglo.Prof(z,100,xlow=200,xup=500)
}
}
