% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rk4d.rl.R
\name{rk4d.rl}
\alias{rk4d.rl}
\title{return level for k4d}
\usage{
rk4d.rl(z, year = c(20, 50, 100, 200), show = F)
}
\arguments{
\item{z}{An object returned by \code{rk4dfit}. The object should represent a stationary model.}

\item{year}{The return level (i.e.\ the profile likelihood is for the value that is exceeded with probability 1/\code{m}).}

\item{show}{Logical; if \code{TRUE} (the default), print details of the result.}
}
\value{
The return level and numeric vector giving the standard errors for the return level.
}
\description{
return level for k4d
}
\examples{
\dontrun{
data(bangkok)
result<-rk4d.fit(bangkok)
rk4d.rl(result$mle,result$cov,100)
}
}
